Refactoring: Split CLI module into two files
